+++ relentless-grpc/examples/service.rs
@@ -4,7 +4,7 @@ use std::process::ExitCode;
 #[cfg(all(feature = "yaml", feature = "console-report"))]
 async fn main() -> Result<ExitCode, Box<dyn std::error::Error + Send + Sync>> {
     use relentless::interface::command::{Assault, Relentless};
-    use relentless_grpc::{client2::DefaultGrpcClient, command::GrpcAssault};
+    use relentless_grpc::{client::DefaultGrpcClient, command::GrpcAssault};
     use relentless_grpc_dev_server::service::{
         counter::{pb::counter_server::CounterServer, CounterImpl},
         echo::{pb::echo_server::EchoServer, EchoImpl},
