+++ relentless-grpc/src/lib.rs
@@ -57,7 +57,6 @@
 //! ```
 
 pub mod client;
-pub mod client2;
 pub mod command;
 pub mod error;
 pub mod evaluate;
